+++ androidApp/src/main/java/com/yaeyama/linerchecker/ui/dashboard/DashBoardPage.kt
@@ -5,17 +5,13 @@ import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.requiredHeight
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Card
 import androidx.compose.material.Divider
-import androidx.compose.material.Icon
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
 import androidx.compose.material.Text
@@ -24,11 +20,10 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.colorResource
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.ikmr.banbara23.yaeyama_liner_checker.R
+import com.yaeyama.linerchecker.ui.dashboard.component.DashBoardHeader
 import com.yaeyama.linerchecker.ui.theme.StatusColor
 import com.yaeyama.linerchecker.ui.theme.YaimafuniAndroidTheme
 import com.yaeyama_liner_checker.domain.statusdetail.Status
@@ -62,28 +57,6 @@ fun RowDivider() {
             .width(1.dp),
         color = colorResource(id = R.color.light_grey),
     )
-}
-
-@Preview(name = "ヘッダー")
-@Composable
-fun DashBoardHeader() {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .requiredHeight(48.dp),
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Spacer(modifier = Modifier.padding(8.dp))
-        Icon(
-            modifier = Modifier.size(24.dp),
-            painter = painterResource(id = R.drawable.ic_harbor),
-            contentDescription = null,
-        )
-        Spacer(modifier = Modifier.padding(8.dp))
-        Text(text = stringResource(id = R.string.dash_board_title))
-
-        // TODO: ProgressBarを配置する（dash_board_fragment.xml の id/port_progressbar を参照）
-    }
 }
 
 @Preview
